requirements.txt
fastapi
sqlalchemy
uvicorn
python-multipart
pyTelegramBotAPI
requests
apscheduler
jinja2
